# healy-apigateway/cmd/docs/swagger.yaml
definitions:
...
  models.DoctorDetails:
    properties:
      email:
        type: string
      fees:
        type: integer
      fullName:
        type: string
      phoneNumber:
        type: string
      specialization:
        type: string
      yearsOfExperience:
        type: integer
    type: object
  models.DoctorLogin:
    description: Doctor login details
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.DoctorSignUp:
    description: Doctor sign up details
    properties:
      confirm_password:
        maxLength: 20
        minLength: 6
        type: string
      email:
        type: string
      fees:
        minimum: 3
        type: integer
      full_name:
        maxLength: 50
        minLength: 3
        type: string
      license_number:
        maxLength: 20
        minLength: 6
        type: string
      password:
        maxLength: 20
        minLength: 6
        type: string
      phone_number:
        maxLength: 15
        minLength: 10
        type: string
      specialization:
        type: string
      years_of_experience:
        maximum: 50
        minimum: 0
        type: integer
    required:
    - confirm_password
    - email
    - fees
    - full_name
    - license_number
    - password
    - phone_number
    - specialization
    - years_of_experience
    type: object
...
  models.Rate:
    properties:
      rate:
        maximum: 5
        minimum: 0
        type: integer
    required:
    - rate
    type: object
...
  response.Response:
    properties:
      data:
        description: |-
          The response data
          example: {"key": "value"}
      error:
        description: |-
          The error information, if any
          example: "Error description"
      message:
        description: |-
          The response message
          example: "Request was successful"
        type: string
    type: object
info:
  contact:
    email: support@healy-medicare.com
    name: API Support
  description: This is the API documentation for the Healy Medicare application.
  title: Healy Medicare API
  version: 1.0.0
paths:
  /admin/dashboard/doctors:
    get:
      consumes:
      - application/json
      description: |-
        List all doctors
        Retrieve the details of all doctors
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      - Bearer: []
      summary: Get all doctors' details
      tags:
      - Admin
      - Patient
...
  /doctor/login:
    post:
      consumes:
      - application/json
      description: Log in an existing doctor
      parameters:
      - description: Doctor Login Details
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.DoctorLogin'
      - description: Doctor Login Details
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.DoctorLogin'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Doctor Login
      tags:
      - Doctor
...
  /doctor/profile:
    get:
      consumes:
      - application/json
      description: Retrieve doctor profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get Doctor Profile
      tags:
      - Doctor
    put:
      consumes:
      - application/json
      description: Update doctor profile
      parameters:
      - description: Doctor Details
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/models.DoctorDetails'
      - description: Doctor Details
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/models.DoctorDetails'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update Doctor Profile
      tags:
      - Doctor
...
  /doctor/signup:
    post:
      consumes:
      - application/json
      description: Sign up a new doctor
      parameters:
      - description: Doctor Sign Up Details
        in: body
        name: doctor
        required: true
        schema:
          $ref: '#/definitions/models.DoctorSignUp'
      - description: Doctor Sign Up Details
        in: body
        name: doctor
        required: true
        schema:
          $ref: '#/definitions/models.DoctorSignUp'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Doctor Sign Up
      tags:
      - Doctor
...
  /patient/doctor:
    get:
      consumes:
      - application/json
      description: |-
        List all doctors
        Retrieve the details of all doctors
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      - Bearer: []
      summary: Get all doctors' details
      tags:
      - Admin
      - Patient
  /patient/doctor/{doctor_id}:
    get:
      consumes:
      - application/json
      description: Retrieve the details of a specific doctor by ID
      parameters:
      - description: Doctor ID
        in: path
        name: doctor_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get individual doctor's details
      tags:
      - Patient
...
  /patient/doctor/rate/{doctor_id}:
    post:
      consumes:
      - application/json
      description: Rate a specific doctor
      parameters:
      - description: Doctor ID
        in: path
        name: doctor_id
        required: true
        type: string
      - description: Rate
        in: body
        name: rate
        required: true
        schema:
          $ref: '#/definitions/models.Rate'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Rate a doctor
      tags:
      - Patient
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"